watchdog timeout extended from 8bit to 16 bit

+++ AVR_Watchdog.cpp
@@ -8,13 +8,13 @@
 #include <avr/wdt.h>
 #include <AVR_Watchdog.h>
 
-uint8_t wdt_count;
-uint8_t wdt_count_limit;
+uint16_t wdt_count;
+uint16_t wdt_count_limit;
 
 void(* resetFunc) (void) = 0;
 
 
-AVR_Watchdog::AVR_Watchdog(uint8_t wdt_seconds)
+AVR_Watchdog::AVR_Watchdog(uint16_t wdt_seconds)
 {
     wdt_count = 0;
     wdt_count_limit = wdt_seconds;   // default value
@@ -27,7 +27,7 @@ void AVR_Watchdog::clear(void)
     wdt_count = 0;
 }
 
-void AVR_Watchdog::set_timeout(uint8_t wdt_seconds)
+void AVR_Watchdog::set_timeout(uint16_t wdt_seconds)
 {
     wdt_count = 0;
     wdt_count_limit = wdt_seconds;
